Add selection volume on tooltip and extract hooks
